Object: Define dummy MSpec methods for compatibility
MSpec defines multiple methods on Object. They are likely to appear in the
files we parse, so we don't want Kernel.load to raise an exception if it
encounters them.

One approach is to use method_missing on Object, but that's particularly ugly.

Another is to rescue NoMethodError after loading each file, but if we did this
unilaterally it may hide bugs. If we were to use a whitelist we'd need to
parse the exception's message to determine what it was raised for, which as
well as being hackish also wouldn't work with alternative implementations
that formatter their exceptions differently.

The current approach is to hard-code a list of MSpec methods, then define each
unless it's already defined. This isn't particularly satisfactory, but will do
for now.